src/ingest/requirements.txt
xarray
astropy
fabio
h5py
pydantic==1.9.0
pymongo
python-dotenv
requests
numpy
Pillow
pandas
git+https://github.com/SciCatProject/pyscicat.git@v0.2.6
typer
matplotlib
pyyaml
pdf2image
python-dateutil
pytz